fix: merge useVisualDiffToggle into useDraftVisualDiff to save 1 import line
Consolidates into single file, removing one import statement from
index.tsx to stay within max-lines budget.
