+++ web_src/src/pages/workflowv2/useDraftVisualDiff.ts
@@ -8,7 +8,7 @@ import type {
 } from "@/api-client";
 import type { CanvasEdge, CanvasNode } from "@/ui/CanvasPage";
 import type { QueryClient } from "@tanstack/react-query";
-import { useMemo } from "react";
+import { useCallback, useMemo, useState } from "react";
 import { buildDraftDiffMap } from "./draftNodeDiff";
 import { prepareNode } from "./workflowPageHelpers";
 
@@ -184,4 +184,23 @@ export function useDraftVisualDiff({
   }, [enabled, isViewingDraftVersion, liveCanvasVersion, preparedEdges]);
 
   return { nodes, edges };
+}
+
+const STORAGE_KEY = "visual-diff-enabled";
+
+export function useVisualDiffToggle() {
+  const [enabled, setEnabled] = useState(() => {
+    const stored = localStorage.getItem(STORAGE_KEY);
+    return stored === null ? true : stored === "true";
+  });
+
+  const toggle = useCallback(() => {
+    setEnabled((prev) => {
+      const next = !prev;
+      localStorage.setItem(STORAGE_KEY, String(next));
+      return next;
+    });
+  }, []);
+
+  return { visualDiffEnabled: enabled, toggleVisualDiff: toggle };
 }
